web/requirements.txt
Django>=3.0
psycopg2-binary
djangorestframework
drf-yasg[validation]